requirements.txt
blinker==1.7.0
boto3==1.34.58
botocore==1.34.58
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-OAuthlib==0.9.6
Flask-Session==0.7.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
msgspec==0.18.6
oauthlib==2.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.18.1
